add a module to get API keys
As I'm going to use more than one API for different actions,
maybe even different weather APIs, I added a modular function
that will return the API key for a specific service. I'm making
this a modular function so I can add more keys in the future
without the need to modify the original function, and call them
in a comfortable way (for example, apiKey.openWeatherMap() for
the openweathermap.org key).